+++ src/CheckLine/CheckLine.css
@@ -1,0 +1,30 @@
+.check-line {
+	display: flex;
+	justify-content:space-between;
+}
+
+.buttons-block {
+	width: 70px;
+	display: flex;
+	justify-content: space-between;
+}
+
+.button-done {
+	margin-right: 0px;
+}
+
+.what-we-need-todo{
+	color: rgb(53, 53, 240);
+	cursor: pointer;
+	/* убирает выделение текста */
+	user-select: none; 
+}
+
+.deal-button{
+	border-radius: 10px;
+	border-color: rgb(111, 111, 241);
+}
+
+.done {
+	text-decoration: line-through;
+}
