+++ src/index.js
@@ -1,4 +1,4 @@
 import React from "react"; 
 import ReactDOM from "react-dom"; 
-import App from './App/App.js';
+import App from './App/App.jsx';
 ReactDOM.render(<App />, document.getElementById("root"));
